probably gonna have to revert this one
added Q and R tuning parameters to pathplanner autobuilder, which is technically some control theory that is over my head. we may be able to use them to help get rid of the oscillations. changed drive feedforward based on theoretical calculations, and added acceleration calculation to left and right drivetrain velocity controllers.